requirements.txt
anyio==4.10.0
beautifulsoup4==4.13.4
black==25.1.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
Events==0.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
mypy_extensions==1.1.0
opensearch-py==3.0.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.3
requests==2.32.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
typing_extensions==4.14.1
urllib3==2.5.0
